move liquid modules to ext namespace
