+++ src/embit/ext/liquid/blech32.py
